# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "llm-libre",
    "summary": "One OpenAI-compatible endpoint for several providers' free LLM models.",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/models": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "List the normalized model catalog",
        "description": "The OpenAI-shaped catalog (`GET /models` format, so OpenAI SDK tooling that lists models works unmodified) plus the gateway's own `auto*` aliases. `owned_by` is the provider id for a real route, or `\"llm-libre\"` for an alias. In production this typically returns 39 entries: the real discovered/declared routes (34) plus the 5 `auto*` aliases -- cross-check the route count against `GET /health`'s `rutas_activas`.",
        "operationId": "modelos_v1_models_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The catalog.",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "object": "list",
                  "data": [
                    {
                      "id": "gpt-5-3-mini",
                      "object": "model",
                      "owned_by": "chatgpt"
                    },
                    {
                      "id": "cohere/north-mini-code:free",
                      "object": "model",
                      "owned_by": "kilo"
                    },
                    {
                      "id": "nvidia/nemotron-3-super-120b-a12b:free",
                      "object": "model",
                      "owned_by": "openrouter"
                    },
                    {
                      "id": "MiniMax-M3",
                      "object": "model",
                      "owned_by": "minimax"
                    },
                    {
                      "id": "auto",
                      "object": "model",
                      "owned_by": "llm-libre"
                    },
                    {
                      "id": "auto:rapido",
                      "object": "model",
                      "owned_by": "llm-libre"
                    },
                    {
                      "id": "auto:potente",
                      "object": "model",
                      "owned_by": "llm-libre"
                    },
                    {
                      "id": "auto:tools",
                      "object": "model",
                      "owned_by": "llm-libre"
                    },
                    {
                      "id": "auto:vision",
                      "object": "model",
                      "owned_by": "llm-libre"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/ranking": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Per-route ranking (not part of the OpenAI contract) -- the operator's main tool to explain a routing decision",
        "description": "One row per active route, with every component that feeds its score broken out, **sorted with the exact same key the router uses** (`(en_cooldown, tier == \"pago\", prioridad, unmeasured, -puntaje)`), not by score alone -- so the top row here is genuinely the route the router would try first right now, matching `X-Ruta-Usada` on the next request. A route currently in cooldown sorts to the bottom of the table even with a perfect score, because the router would not pick it at this instant either.\n\n`calidad` is `null` and `calidad_asumida` holds the neutral value instead whenever `calidad_medida` is `false`: a route that has never been through the quality battery is not silently shown as if a 0.6 had been measured.\n\n`en_cooldown_hasta` is a **raw Unix timestamp in seconds** (`0` means not in cooldown) -- unlike `ultima_sonda` / `ultima_sonda_calidad`, which are ISO-8601 strings (or `null`).\n\n`ttft_p50_ms` (time to first token) is only ever measured by streaming traffic and probes; `latencia_p50_ms` is the full round-trip and does not feed the score, only diagnostics -- the two are different magnitudes, do not average them together.",
        "operationId": "ranking_v1_ranking_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The ranking table.",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "rutas": [
                    {
                      "clave": "chatgpt/gpt-5-3-mini",
                      "tier": "gratis",
                      "prioridad": 0,
                      "puntaje": 0.7912,
                      "calidad": 0.8,
                      "calidad_medida": true,
                      "ultima_sonda_calidad": "2026-08-17T03:12:04Z",
                      "ultima_sonda": "2026-08-17T08:00:11Z",
                      "confiabilidad": 0.98,
                      "ttft_p50_ms": 900.0,
                      "latencia_p50_ms": 4500.0,
                      "en_cooldown_hasta": 0.0,
                      "tools": false,
                      "vision": false,
                      "contexto": 128000
                    },
                    {
                      "clave": "kilo/cohere/north-mini-code:free",
                      "tier": "gratis",
                      "prioridad": 1,
                      "puntaje": 0.842,
                      "calidad": 0.8,
                      "calidad_medida": true,
                      "ultima_sonda_calidad": "2026-08-17T03:14:51Z",
                      "ultima_sonda": "2026-08-17T08:01:02Z",
                      "confiabilidad": 1.0,
                      "ttft_p50_ms": 650.0,
                      "latencia_p50_ms": 2100.0,
                      "en_cooldown_hasta": 0.0,
                      "tools": true,
                      "vision": false,
                      "contexto": 128000
                    },
                    {
                      "clave": "openrouter/poolside/laguna-s-2.1:free",
                      "tier": "gratis",
                      "prioridad": 1,
                      "puntaje": 0.5015,
                      "calidad": 0.6,
                      "calidad_medida": false,
                      "calidad_asumida": 0.6,
                      "ultima_sonda": "2026-08-17T08:02:00Z",
                      "confiabilidad": 0.8,
                      "ttft_p50_ms": 1500.0,
                      "en_cooldown_hasta": 1755400930.0,
                      "tools": false,
                      "vision": false,
                      "contexto": 65536
                    },
                    {
                      "clave": "minimax/MiniMax-M3",
                      "tier": "pago",
                      "prioridad": 2,
                      "puntaje": 0.95,
                      "calidad": 0.9,
                      "calidad_medida": true,
                      "confiabilidad": 0.8,
                      "ttft_p50_ms": 1500.0,
                      "en_cooldown_hasta": 0.0,
                      "tools": true,
                      "vision": false,
                      "contexto": 128000
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/uso": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Today's paid-usage counter for the calling key",
        "description": "How many requests the calling key spent against the paid fallback (MiniMax) today (UTC), against its daily allowance (`TOPE_PAGO_DIARIO`). Counts every **billable** attempt, not only successful ones -- a `200` from the paid provider is billable whether or not it contained a usable answer, and streaming counts exactly once per request, not once per chunk. This is the only place a streaming client can see whether the paid fallback served (or was attempted for) its last request, since streaming responses do not carry `X-Tier`.",
        "operationId": "uso_v1_uso_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Today's paid usage for this key.",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "dia": "2026-08-17",
                  "pago_hoy": 0,
                  "tope": 200
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "One of the configured LLM_LIBRE_API_KEYS. Wins over Authorization if both headers are present."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Same keys as ApiKeyHeader, via the header any OpenAI SDK sends automatically for its `api_key` parameter."
      }
    }
  }
}